backend/middleware/auth.py
"""
JWT Authentication Middleware
Handles JWT token verification and user authentication
"""

import os
from datetime import datetime, timedelta
from typing import Optional
from fastapi import Request, HTTPException, status
from starlette.middleware.base import BaseHTTPMiddleware
import jwt

# Get JWT secret from environment
JWT_SECRET_KEY = os.getenv("JWT_SECRET_KEY", "your-secret-key-change-in-production")
ALGORITHM = "HS256"
ACCESS_TOKEN_EXPIRE_MINUTES = 30


class JWTMiddleware(BaseHTTPMiddleware):
    """
    JWT authentication middleware
    TODO: Implement token verification and user context injection
    
    NOTE: This middleware is disabled by default in main.py until fully implemented.
    To use, uncomment in main.py and complete the implementation below.
    """

    async def dispatch(self, request: Request, call_next):
        """
        Process request and verify JWT token
        TODO: Extract and validate JWT from Authorization header
        TODO: Inject user context into request state
        """
        # Skip token verification for public endpoints
        public_endpoints = ["/health", "/api/auth/register", "/api/auth/login", "/docs", "/openapi.json"]
        
        if request.url.path in public_endpoints or any(request.url.path.startswith(ep) for ep in public_endpoints):
            return await call_next(request)

        # TODO: Implement JWT verification
        # - Extract token from Authorization header
        # - Verify token signature and expiration
        # - Extract user information from token
        # - Store in request.state for route handlers

        response = await call_next(request)
        return response


def create_access_token(data: dict, expires_delta: Optional[timedelta] = None) -> str:
    """
    Create JWT access token
    TODO: Implement token generation with expiration
    
    Args:
        data: Dictionary containing token claims
        expires_delta: Optional expiration delta
    
    Returns:
        str: Encoded JWT token
    """
    pass


def verify_token(token: str) -> dict:
    """
    Verify JWT token and return payload
    TODO: Implement token verification with error handling
    
    Args:
        token: JWT token string
    
    Returns:
        dict: Decoded token payload
    
    Raises:
        HTTPException: If token is invalid or expired
    """
    pass


def hash_password(password: str) -> str:
    """
    Hash password using bcrypt
    TODO: Implement bcrypt password hashing
    
    Args:
        password: Plain text password
    
    Returns:
        str: Hashed password
    """
    pass


def verify_password(plain_password: str, hashed_password: str) -> bool:
    """
    Verify password against hash
    TODO: Implement bcrypt password verification
    
    Args:
        plain_password: Plain text password
        hashed_password: Hashed password from database
    
    Returns:
        bool: True if password matches
    """
    pass
