backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Cache
redis==5.0.1

# Authentication & Security
pyjwt==2.12.1
bcrypt==4.1.1
python-multipart==0.0.6

# API Clients
google-generativeai==0.3.0  # Gemini API
groq==0.4.1  # Groq API

# Environment & Configuration
python-dotenv==1.0.0

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
requests==2.31.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Logging & Monitoring
python-json-logger==2.0.7
